fix empty name-tag text in 1.18
1.18's Text#copy method (more precisely LiteralText#copy)
only copies the String content of the Text, and not the
siblings and style of the Text, which results in an empty
name-tag when we try to render the pop counter.

To address this issue, instead of copying the text, we
check if it is mutable, and it is, use the MutableText
itself instead of the copy, preserving the style and
siblings.